refactor: refactoring authentication api
