--- src/main/java/kr/or/cola/backend/mail/MailService.java
@@ -1,19 +1,0 @@
-package kr.or.cola.backend.mail;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-
-public class MailService {
-    private final JavaMailSender javaMailSender;
-
-    @Async
-    public void sendEmail(SimpleMailMessage email) {
-        javaMailSender.send(email);
-    }
-}
