+++ src/main/java/kr/or/cola/backend/util/CustomMailSender.java
@@ -1,0 +1,27 @@
+package kr.or.cola.backend.util;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class CustomMailSender {
+    private final JavaMailSender javaMailSender;
+
+    public void sendAuthenticationToken(String authToken, String receiverEmail) {
+        SimpleMailMessage mailMessage = new SimpleMailMessage();
+        mailMessage.setTo(receiverEmail);
+        mailMessage.setSubject("회원가입 이메일 인증");
+        mailMessage.setText(authToken);
+
+        sendEmail(mailMessage);
+    }
+
+    @Async
+    public void sendEmail(SimpleMailMessage email) {
+        javaMailSender.send(email);
+    }
+}
